+++ crate/src/pipeline/mod.rs
@@ -1,0 +1,9 @@
+mod builder;
+mod ops;
+mod planar;
+
+pub use builder::{
+    pipeline_conversion_roundtrip, pipeline_invert, pipeline_invert_alter_channels,
+    Pipeline,
+};
+pub use planar::PlanarImage;
