+++ crate/src/pipeline/planar.rs
@@ -1,14 +1,16 @@
 use crate::PhotonImage;
 use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "enable_wasm")]
 use wasm_bindgen::prelude::wasm_bindgen;
 
 #[cfg_attr(feature = "enable_wasm", wasm_bindgen)]
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct PlanarImage {
-    r: Vec<u8>,
-    g: Vec<u8>,
-    b: Vec<u8>,
-    a: Vec<u8>,
+    pub(super) r: Vec<u8>,
+    pub(super) g: Vec<u8>,
+    pub(super) b: Vec<u8>,
+    pub(super) a: Vec<u8>,
     width: u32,
     height: u32,
 }
@@ -243,280 +245,5 @@ impl PlanarImage {
             width: self.width,
             height: self.height,
         }
-    }
-}
-
-pub struct Pipeline {
-    image: PlanarImage,
-    pending: Vec<PixelOp>,
-}
-
-impl Pipeline {
-    pub fn from_photon_image(img: &PhotonImage) -> Self {
-        let planar_image = PlanarImage::from_photon_image(img);
-        Pipeline {
-            image: planar_image,
-            pending: Vec::new(),
-        }
-    }
-
-    pub fn finish(mut self) -> PhotonImage {
-        self.flush_pixel_ops();
-        self.image.to_photon_image()
-    }
-
-    pub fn gray_scale(mut self) -> Self {
-        self.pending.push(PixelOp::GrayScale);
-        self
-    }
-
-    pub fn monochrome(mut self, r_offset: u8, g_offset: u8, b_offset: u8) -> Self {
-        self.pending.push(PixelOp::Monochrome {
-            r_offset,
-            g_offset,
-            b_offset,
-        });
-        self
-    }
-
-    pub fn invert(mut self) -> Self {
-        self.pending.push(PixelOp::Invert);
-        self
-    }
-
-    pub fn alter_channels(mut self, r: i16, g: i16, b: i16) -> Self {
-        self.pending.push(PixelOp::AlterChannels { r, g, b });
-        self
-    }
-
-    pub fn swap_channels(mut self, mut channel1: usize, mut channel2: usize) -> Self {
-        if channel1 > 2 {
-            panic!(
-                "Invalid channel index passed. Channel1 must be equal to 0, 1, or 2."
-            );
-        }
-        if channel2 > 2 {
-            panic!(
-                "Invalid channel index passed. Channel2 must be equal to 0, 1, or 2."
-            );
-        }
-
-        self.flush_pixel_ops();
-
-        if channel1 == channel2 {
-            return self;
-        }
-
-        if channel1 > channel2 {
-            std::mem::swap(&mut channel1, &mut channel2);
-        }
-
-        match (channel1, channel2) {
-            (0, 1) => std::mem::swap(&mut self.image.r, &mut self.image.g),
-            (0, 2) => std::mem::swap(&mut self.image.r, &mut self.image.b),
-            (1, 2) => std::mem::swap(&mut self.image.g, &mut self.image.b),
-            _ => unreachable!(),
-        }
-
-        self
-    }
-
-    fn flush_pixel_ops(&mut self) {
-        if self.pending.is_empty() {
-            return;
-        }
-
-        #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-        unsafe {
-            self.flush_pixel_ops_simd();
-        }
-
-        #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
-        {
-            self.flush_pixel_ops_scalar();
-        }
-    }
-
-    #[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
-    fn flush_pixel_ops_scalar(&mut self) {
-        for i in 0..self.image.r.len() {
-            let mut r = self.image.r[i];
-            let mut g = self.image.g[i];
-            let mut b = self.image.b[i];
-
-            for op in &self.pending {
-                apply_pixel_op_scalar(op, &mut r, &mut g, &mut b);
-            }
-
-            self.image.r[i] = r;
-            self.image.g[i] = g;
-            self.image.b[i] = b;
-        }
-
-        self.pending.clear();
-    }
-
-    #[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-    #[target_feature(enable = "simd128")]
-    unsafe fn flush_pixel_ops_simd(&mut self) {
-        for op in &self.pending {
-            match op {
-                PixelOp::Invert => invert_planes_simd(
-                    &mut self.image.r,
-                    &mut self.image.g,
-                    &mut self.image.b,
-                ),
-                PixelOp::AlterChannels { r, g, b } => alter_channels_planes_simd(
-                    &mut self.image.r,
-                    &mut self.image.g,
-                    &mut self.image.b,
-                    *r,
-                    *g,
-                    *b,
-                ),
-                _ => todo!(
-                    "There is no SIMD-optimized implementation for this operation yet"
-                ),
-            }
-        }
-
-        self.pending.clear();
-    }
-}
-
-#[cfg_attr(feature = "enable_wasm", wasm_bindgen)]
-pub fn pipeline_conversion_roundtrip(img: &mut PhotonImage) {
-    let output = Pipeline::from_photon_image(img).finish();
-    img.raw_pixels = output.raw_pixels;
-}
-
-#[cfg_attr(feature = "enable_wasm", wasm_bindgen)]
-pub fn pipeline_invert(img: &mut PhotonImage) {
-    let output = Pipeline::from_photon_image(img).invert().finish();
-    img.raw_pixels = output.raw_pixels;
-}
-
-#[cfg_attr(feature = "enable_wasm", wasm_bindgen)]
-pub fn pipeline_invert_alter_channels(img: &mut PhotonImage, r: i16, g: i16, b: i16) {
-    let output = Pipeline::from_photon_image(img)
-        .invert()
-        .alter_channels(r, g, b)
-        .finish();
-    img.raw_pixels = output.raw_pixels;
-}
-
-#[allow(dead_code)]
-enum PixelOp {
-    GrayScale,
-    Monochrome {
-        r_offset: u8,
-        g_offset: u8,
-        b_offset: u8,
-    },
-    Invert,
-    AlterChannels {
-        r: i16,
-        g: i16,
-        b: i16,
-    },
-}
-
-#[cfg(not(all(target_arch = "wasm32", target_feature = "simd128")))]
-fn apply_pixel_op_scalar(op: &PixelOp, r: &mut u8, g: &mut u8, b: &mut u8) {
-    match op {
-        PixelOp::GrayScale => {
-            let avg = ((*r as u32 + *g as u32 + *b as u32) / 3) as u8;
-            *r = avg;
-            *g = avg;
-            *b = avg;
-        }
-        PixelOp::Monochrome {
-            r_offset,
-            g_offset,
-            b_offset,
-        } => {
-            let avg = (*r as u32 + *g as u32 + *b as u32) / 3;
-            *r = (avg + *r_offset as u32).min(255) as u8;
-            *g = (avg + *g_offset as u32).min(255) as u8;
-            *b = (avg + *b_offset as u32).min(255) as u8;
-        }
-        PixelOp::Invert => {
-            *r = 255 - *r;
-            *g = 255 - *g;
-            *b = 255 - *b;
-        }
-        PixelOp::AlterChannels {
-            r: dr,
-            g: dg,
-            b: db,
-        } => {
-            *r = (*r as i16 + dr).clamp(0, 255) as u8;
-            *g = (*g as i16 + dg).clamp(0, 255) as u8;
-            *b = (*b as i16 + db).clamp(0, 255) as u8;
-        }
-    }
-}
-
-#[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-#[target_feature(enable = "simd128")]
-unsafe fn invert_planes_simd(r: &mut [u8], g: &mut [u8], b: &mut [u8]) {
-    use core::arch::wasm32::*;
-
-    let simd_len = r.len() - (r.len() % 16);
-    let mask = u8x16_splat(255);
-
-    for i in (0..simd_len).step_by(16) {
-        let rv = v128_load(r.as_ptr().add(i) as *const v128);
-        let gv = v128_load(g.as_ptr().add(i) as *const v128);
-        let bv = v128_load(b.as_ptr().add(i) as *const v128);
-
-        v128_store(r.as_mut_ptr().add(i) as *mut v128, v128_xor(rv, mask));
-        v128_store(g.as_mut_ptr().add(i) as *mut v128, v128_xor(gv, mask));
-        v128_store(b.as_mut_ptr().add(i) as *mut v128, v128_xor(bv, mask));
-    }
-
-    for i in simd_len..r.len() {
-        r[i] = 255 - r[i];
-        g[i] = 255 - g[i];
-        b[i] = 255 - b[i];
-    }
-}
-
-#[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-#[target_feature(enable = "simd128")]
-unsafe fn alter_channels_planes_simd(
-    r: &mut [u8],
-    g: &mut [u8],
-    b: &mut [u8],
-    r_amt: i16,
-    g_amt: i16,
-    b_amt: i16,
-) {
-    alter_plane_simd(r, r_amt);
-    alter_plane_simd(g, g_amt);
-    alter_plane_simd(b, b_amt);
-}
-
-#[cfg(all(target_arch = "wasm32", target_feature = "simd128"))]
-#[target_feature(enable = "simd128")]
-unsafe fn alter_plane_simd(channel: &mut [u8], amt: i16) {
-    use core::arch::wasm32::*;
-
-    let simd_len = channel.len() - (channel.len() % 16);
-    let magnitude = amt.unsigned_abs().min(255) as u8;
-    let delta = u8x16_splat(magnitude);
-
-    for i in (0..simd_len).step_by(16) {
-        let values = v128_load(channel.as_ptr().add(i) as *const v128);
-        let output = if amt >= 0 {
-            u8x16_add_sat(values, delta)
-        } else {
-            u8x16_sub_sat(values, delta)
-        };
-        v128_store(channel.as_mut_ptr().add(i) as *mut v128, output);
-    }
-
-    for value in &mut channel[simd_len..] {
-        *value = (*value as i16 + amt).clamp(0, 255) as u8;
     }
 }
